Add a setting for the timeout stack style:
"Stack" is the default behaviour, it will search 20 messages up and 5
seconds back in time to stack the timeout.
"Stack sparingly" will try to do the same, but only if the user has not
typed a message inbetween the this and the last timeout.

Fixes #1157

This commit has been anonymized with Repush.