Highlight parsing is now done after Ignores have been respected.
In the same commit, because I misunderstood the original issue, I also
split out the parsing of the highlights and triggering of the highlight
sounds/alerts into separate functions.

Fixes #1160

I also re-enabled the "sound url changer" that was commented out in an
earlier commit.

Fixes #1170

This commit has been anonymized with Repush.