Fix "Manage Accounts" button
This commit fixes the unresponsiveness of the "Manage Account" button
(described in #1188).

Apparently, focus was lost when pressing down the mouse button. This is
circumvented by setting the focus policy of the button to `Qt::NoFocus`.
The button then works as expected and opens the "Account" page in the
preferences.

This commit has been anonymized with Repush.