Fix subscription messages triggering split highlights (#1519)
Since #1320, subscription messages are treated as highlights in order to
allow customization. This caused subscription messages to highlight the
split(s) the message was received in. This is not intended behavior.

This commit fixes the issue by additionally checking if the
`Subscription` flag is set on a highlighted message.

This commit has been anonymized with Repush.