Better Highlights: Fix Unintentional Color Update (#1522)
* HighlightPhrase: Fix wrong documentation

* Use right constructor for new HighlightPhrases

* Fix preset highlights changing unintentionally

Prior to this commit, the callback for reacting to user input on the
highlight table (namely, `HighlightingPage::tableCellClicked`) only
checked for the row number in order to determine whether preset
highlights (self highlights, whispers, and subscriptions) need to be
updated. Hence, changing rows 0 through 2 in the "User Highlights" tab
would also update the preset highlights.

This commit adds a check to determine whether the callback was triggered
by the "Messages" highlight tab, or not.

This commit has been anonymized with Repush.