Only attempt to read thumbnails when we get a 200 response code.
We would have been able to keep doing this if the Network Request code
followed redirects - however, it doesn't, so this is the best we can do.
We also don't have header support in Network Result, so we can't
validate that the response we got is actually an image.

Fixes #1972

This commit has been anonymized with Repush.