Fix: ColorButtons in settings don't respect search (#2073)
Prior to this commit, adding a ColorButton to a SettingsLayout via
`SettingsLayout::addColorButton` lead to the button not respecting
search queries. This is because they are not added to the layout in a
Group. This commit fixes the behavior, causing color buttons to behave
like every other settings widget.

This commit has been anonymized with Repush.