Add StartupWMClass property to desktop file (#2587)
This adds the StartupWMClass property to the desktop file, which window managers can use to determine when chatterino has opened and which window belongs to it.

GNOME uses this property, so this will fix #1863. You can test this by applying this patch to the desktop file of an existing install.

Relevant FreeDesktop docs: specifications.freedesktop.org/desktop-entry-spec/latest/ar01s06.html#key-startupwmclass

Fixes #1863

This commit has been anonymized with Repush.